backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
langchain==0.0.335
langchain-openai==0.0.2
chromadb==0.4.18
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-multipart==0.0.6
